requirements.txt
pyinstaller==5.10.0
Pillow==9.5.0
bs4==0.0.1
requests==2.28.2
requests-html==0.10.0
